pyproject.toml
[tool.poetry]
name = "qa-guru-diplom-mob"
version = "0.1.0"
description = ""
authors = ["eisaeva <hanymka@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.2.0"
allure-pytest = "^2.12.0"
selene = "^2.0.0-b12"
Appium-Python-Client = "^2.6.1"
python-dotenv = "^0.21.0"
config = "^0.5.1"

[tool.pytest.ini_options]
addopts = "--alluredir allure-results --clean-alluredir"
filterwarnings = [
    "error",
    "ignore::UserWarning",
    'ignore: HTTPResponse.getheader\(\) is deprecated:DeprecationWarning'
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"